touch: goodix_9916r: Add parameter check to avoid security issue
In brl_read_config function, it is possible the parameter size
is smaller than cfg_head, and if in some case the cfg is corrupted,
the code will possible reach memcpy function, in this function,
we might be writing to a place where cannot even hold 64 bytes.
That will cause security issue.

Change-Id: I6ac69de62e8565dc4bee4befbe3124c60118f3a1
Signed-off-by: Yu Wu <quic_zwy@quicinc.com>
Signed-off-by: Fiqri Ardyansyah <fiqri191002@gmail.com>