input: touchscreen: goodix_berlin_driver: Add support to disable rawdata mode on thp enabled devices
Some devices ship a firmware which enables host based touch
processing (proprietary userspace binary). Since we don't
want to use that but still use the stock firmware, we need
to disable the rawdata mode and enable the coor mode on init
and each resume. To enable this, add the goodix,support-thp-fw
property to devicetree.

Change-Id: I55c606a6c468663cadc9506cb7d08e5cf2965bc7
Signed-off-by: Fiqri Ardyansyah <fiqri191002@gmail.com>